test: 프론트 웹소켓 테스트 수정
