+++ src/App.tsx
@@ -1,10 +1,14 @@
-import WebSocketText from "pages/WebSocketText";
 import React from "react";
-// import { Route, Routes } from "react-router-dom";
-// import PlaygroudPage from "./pages/PlaygroudPage";
+import { Route, Routes } from "react-router-dom";
+import PlaygroudPage from "./pages/PlaygroudPage";
 
 function App() {
-  return <WebSocketText />;
+  return (
+    <Routes>
+      <Route path={"/playground"} element={<PlaygroudPage />} />
+      <Route path={"/"} element={<PlaygroudPage />} />
+    </Routes>
+  );
 }
 
 export default App;
